refactor: move components from stocks to shared directory
