+++ resources/js/components/delete-user.tsx
@@ -1,14 +1,14 @@
 import { useForm } from '@inertiajs/react';
 import { FormEventHandler, useRef } from 'react';
 
-import InputError from '@/components/stocks/input-error';
+import InputError from '@/components/shared/input-error';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
 
 import HeadingSmall from '@/components/heading-small';
 
 import { Dialog, DialogClose, DialogContent, DialogDescription, DialogFooter, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
-import InputShowPassword from '@/components/stocks/input-show-password';
+import InputShowPassword from '@/components/shared/input-show-password';
 
 export default function DeleteUser() {
     const passwordInput = useRef<HTMLInputElement>(null);
