+++ resources/js/components/example/create-example-form.tsx
@@ -1,5 +1,5 @@
-import ButtonSubmit from '@/components/stocks/button-submit';
-import InputError from '@/components/stocks/input-error';
+import ButtonSubmit from '@/components/shared/button-submit';
+import InputError from '@/components/shared/input-error';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { useForm } from '@inertiajs/react';
