+++ resources/js/components/shared/input-strong-password.tsx
