+++ resources/js/pages/auth/confirm-password.tsx
@@ -5,8 +5,8 @@ import { Head, useForm } from '@inertiajs/react';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import AuthLayout from '@/layouts/auth-layout';
-import InputError from '@/components/stocks/input-error';
-import ButtonSubmit from '@/components/stocks/button-submit';
+import InputError from '@/components/shared/input-error';
+import ButtonSubmit from '@/components/shared/button-submit';
 
 export default function ConfirmPassword() {
     const { data, setData, post, processing, errors, reset } = useForm<Required<{ password: string }>>({
