+++ resources/js/pages/auth/forgot-password.tsx
@@ -6,8 +6,8 @@ import TextLink from '@/components/text-link';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import AuthLayout from '@/layouts/auth-layout';
-import InputError from '@/components/stocks/input-error';
-import ButtonSubmit from '@/components/stocks/button-submit';
+import InputError from '@/components/shared/input-error';
+import ButtonSubmit from '@/components/shared/button-submit';
 
 export default function ForgotPassword({ status }: { status?: string }) {
     const { data, setData, post, processing, errors } = useForm<Required<{ email: string }>>({
